requirements.txt
flask
pillow
cryptography
numpy